refactor(core): Add isUniqueConstraintError unit tests and rename deactivator to PublishedWorkflowTriggerDeactivator (no-changelog)
Add a dedicated @n8n/db unit test for isUniqueConstraintError (moving the cases out of the cli response-helper test), and rename PublicationTriggerDeactivator to PublishedWorkflowTriggerDeactivator to match PublishedWorkflowEnqueuer.
